Version 0.6.0 brings mostly "debian compatiblity" as far as i can test it for now.
The project now ships with a setup.rb, in addition to the `gem install GEMNAME` installation-approach. This also installs the man-pages into the right directory. The rdoc-able API-doc and the README have also been updated and extended, where applicable.

An architectural change is the centralised management of direcotries.  Both reader-classes utilise this now. This should clear the way for future enhancements.

I also took the management of the gemspec away from jeweler, which I therefore don't need anymore. I use gem-release for tagging and releasing, but that's mostly a convenience wrapper.