started development after 0.7.0
ToPass now runs seamlessly with Ruby 1.8.7, 1.9.2, REE 2010-02. Rubinius works mostly, but has some issues.

The Metainformation like version, summary, description and the likes are now all in the version.rb-file. This way, they can be queried easier.

The test can now be run through the test/all.rb-script. In fact, this is the standard way and internally used by `rake test`. The Manpages are now in the right location for setup.rb and RDocs now include the LICENSE and this
CHANGELOG.

The gemspec has the filelist now directly from git instead from the filesystem.